+++ CHANGELOG.md
@@ -3,11 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [1.2.X.X][1.2.X.X]
+## [1.2.1.0][1.2.1.0]
 
 ### Added
-*   Example for Flexipay direct.
+*   Parameters `paymentReference` and `invoiceId` to `Authorization` and `Payout`.
 *   The SDK now supports the webhook event `payout`.
+*   Example for Flexipay direct.
 
 ### Changed
 *   The webhook tests now cover all supported events.
@@ -252,3 +253,4 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 [1.1.5.0]: https://github.com/heidelpay/heidelpayPHP/compare/1.1.4.0..1.1.5.0
 [1.1.6.0]: https://github.com/heidelpay/heidelpayPHP/compare/1.1.5.0..1.1.6.0
 [1.2.0.0]: https://github.com/heidelpay/heidelpayPHP/compare/1.1.6.0..1.2.0.0
+[1.2.1.0]: https://github.com/heidelpay/heidelpayPHP/compare/1.2.0.0..1.2.1.0